🔄 Add Regex timeout to validation of email addresses, phone numbers, and urls
The validation attributes for these things in
`System.ComponentModel.DataAnnotations` do not enforce a timeout when
matching their regular expressions. These regex's are "evil"[1] since they
can get stuck in a loop when evaluating certain carefully crafted input.
This leads to a DoS vulnerability on forms that use these attributes for
validation.

This change adds our own implementation of `EmailAttribute`,
`PhoneAttribute`, and `UrlAttribute` which enforce a timeout of one
second when matching input. If it takes longer than one second to match,
a `RegexMatchTimeoutException` will be thrown.

See https://❌🤐🔙.❌🤐🔙.org/cgi-bin/❌🤐🔙name.cgi?name=❌🤐🔙-2015-2526 and
http://blog.malerisch.net/2015/09/net-mvc-redos-denial-of-service-vulnerability-❌🤐🔙-2015-2526.html
for more info.

[1] https://www.owasp.org/index.php/Regular_expression_Denial_of_Service_-_ReDoS


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/d460920142b989e6e86080785733ed545d3aa25b