🔄 Update NuGet package NUnit to 3.0.1
Upgrades NUnit to the next major ❌🤐🔙. This was a breaking change
with several features we previously relied on either marked obsolete
or removed entirely. In the spirit of keeping this commit small and
focused, only the minimal code changes necessary to make the tests
build and pass were made. Of particular note:
    - NUnit 3 removed the ExpectedExceptionAttribute and the
      TestCaseData.Throws method. Tests which used either of these were
      updated to use the Assert.Throws style of checking that an
      exception was thrown. In some cases, tests which used the
      ExpectedException in conjunction with the TestCaseAttribute had to
      be split into multiple tests to make the new assert style work.
      Additionally, Assert.Throws doesn't yet support async methods, so
      a helper was added for those cases.
    - NUnit 3 also removed the Result attribute. Tests which used it
      with a return value were changed to use an assert equal on the
      expected and actual values.
    - The TestFixtureSetup and TestFixtureTearDown attributes were
      marked obsolete in favor of OneTimeSetup and OneTimeTearDown. All
      usages were replaced with the new attributes.
    - TestCaseSource no longer supports instance methods or instance
      fields as a source. All methods and fields referenced in a
      TestCaseSource attribure were made static. In some cases, minor
      refactoring was needed to make this possible. Additional
      refactoring, outside the scope of this commit, should be
      considered in the future.

The ReSharper 9.x test runner and the Visual Studio 2015 test runner do
not yet support NUnit 3 and there is not yet an updated NuGet package for
the NUnit test adapter which supports NUnit 3. So this change will
require that everyone is running ReSharper 10. There is an extension[1]
on the Visual Studio gallery which adds NUnit 3 support to Visual
Studio, but this must be installed and configured by each team member.

[1]https://visualstudiogallery.msdn.microsoft.com/0da0f6bd-9bb6-4ae3-87a8-537788622f2d


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/c542ebe9e03029eaa3923cc8af9650de5bc66087