🔄 Show only available workflow steps on ❌🤐🔙 actions modal
Previously, the server would enforce that users only update the workflow
step of an item to a step which is allowed to follow its current step,
but the modal would list all the steps. This change causes the modal to
only show the steps which are valid from the current step by building
the dropdown from what the server says are allowed. The dropdown will
be repopulated when the step is changed with the new allowed next steps.

This required changing the update call from change on the dropdown
selection to a button click since repopulating the options caused the
change event to fire.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/626536f2b6876b5ad43ac3a8f13b596e0c08fe71