🔄 Split Binding Extensions into Multiple Classes
CLR references are weird. This simple class/namespace split makes it so that assemblies using some of the `.In(…)` extension methods don't require that we reference all the assemblies containing the types on all of the signatures of the `.In(…)` extension methods. That is, if you are using `.In(…)` for a binding that isn't a web filter, you shouldn't need to reference web assemblies on account of you using `.In(…)`.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/31677fe0b3adbde9ba819bc3338856f899b5e0b8