🔄 Fix Rounding Error In Integration Test
Some integration tests will randomly fail due to an improper use of rounding. Previously `DateTimeOffset`s were rounded in the test to check for equality between the object that was sent to the database and the object that was retrieved from the database. Due to MSSQL storing `DateTimeOffset`s with less precision than those stored in C#, the extra ticks (not stored in MSSQL) are dropped and thereby floored. Before we had been attempting to round both the value from the database and the value we put in the database to compensate for this, which was incorrect, the correct behavior is to floor both values just like the database does.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/a60154d328b8062adc77ec77f4b8cde6e6a7ed3c