🔄 Introduced ApiFilterDecoraptor to Control Api Filter Lifetimes
Ninject was re-using instances of our Api filters regardless as to what lifetime we were binding them with, so this change converts existing Api filter bindings (and makes it easier for future bindings) to bind a decoraptor for each filter when the filter is bound. The decoraptor can then defer lifetime initialization of the actual filter until it is needed. All this prevents us from getting problems from trying to "reuse" `PumpDbContext`, and other fun lifetime problems.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/a21e90bb9dd9bd4ad81feabe6819f321ad5fa0a8