🔄 Add metadata from ❌🤐🔙 to output of parsed ❌🤐🔙s
The ❌🤐🔙 provides some metadata around the XML feeds for ❌🤐🔙s, including
file size, checksum, and last modified date. This metadata is available
by replacing the extension of the compressed XML feed with `.meta` (e.g.
https://❌🤐🔙.❌🤐🔙.gov/feeds/xml/❌🤐🔙/❌🤐🔙-2.0-Modified.xml.gz. would
become https://❌🤐🔙.❌🤐🔙.gov/feeds/xml/❌🤐🔙/❌🤐🔙-2.0-Modified.meta). See
the "META files" section of https://❌🤐🔙.❌🤐🔙.gov/download.cfm#❌🤐🔙_FEED

This change adds that metadata, along with a count of the ❌🤐🔙s, to the
output of the parsed ❌🤐🔙s. ❌🤐🔙 provides the last modified date with
an EST/EDT offset. For consistency with the rest of the dates in the
application, this was converted to UTC before being presented.
Additionally the presentation of the fields was cleaned up to be more
friendly to the users: last modified was humanaized and the file size
was converted from bytes to a more appropriate measure (KB/MB/GB etc)
based on the value (this required moving some code out of the Upload
model to a more central location).


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/89880f3b1da05d5fd1a2825882d673b7e98e24c0