🔄 Fix Self-Signed CA errors for ❌🤐🔙
❌🤐🔙 repositories use an in-house Certificate Authority to sign their
TLS certificates for their package repositories. Because the CA
certificate is self-signed and not trusted by our servers, we need to
allow the CA certificate to be trusted.

Installing the CA certificate on the server would make a larger security
hole. Additionally these could be loaded from disk on the server, but if
the certificates on disk were to be replaced by a malicious user we
could also get bad, untrusted data from an unknown source. For this
reason I have hard coded the trusted certificate into the source,
allowing us to ❌🤐🔙 control the certificate and better ensure the
integrity of the certificate.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/6e3b4cd22b96f276f863f9c2aff6cb0133e75bb1