🔄 Refactor Linux Integration
Previously the Linux integration was too slow to handle the amount of
data produced by attempting to import a Red Hat repository. This speeds
it up by way of doing smaller increments of work in each Javascript
request.

The steps are:
    * Dump all the raw xml from both files into the database
    * Extract the relational keys (filenames) from both files and save
      in the database
    * Relate the two entries based on the extracted relational keys,
      parse out the ❌🤐🔙 ❌🤐🔙 ❌🤐🔙 and ❌🤐🔙 entries and insert
      them into the PUMP database.
    * Get all the failed items and display them
    * Delete all the staging information except the Repository Import
      Information (name, links to .xml.gz files)

The unused old code was removed as part of this effort.

Some of the `WebClient` code was reworked to remove some of the data
copies that were previously happening (considering the amount of data
being transferred by this operation).


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/ffec7d965acf7444dc8b17b764bc1a2173e3d2bb