🔄 Add ability to temporarily stop time
Adds an implementation of `IClock` which can be "frozen" to preserve the
next requested "now" time until it is "unfrozen". This allows us to
ensure that we get consistent time values when needed, but in a
decoupled fashion. The goal is for consumers not to know that they are
interacting with a frozen clock.

This was accomplished by adding an interceptor and the `FreezeTime`
attribute. When this attribute is applied to a method, time will be
frozen before it is invoked and unfrozem immediately after.

Any classes that do not want to receive frozen time need to
be registered with the system clock explicitly.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/87a6cd21a8a5715166b223d52260ad37956498a8