🔄 Setup retry policy in FGS.Pump.FaultHandling using Polly
Introduces a general purpose retry policy implemented using polly. This
policy is essentially an adapter over the polly retry policy that is
customized to our needs. (Polly has a bunch of stuff exposed on their
policies we don't need and it is also fond of being configured statically
which is antithetical to our DI configuration and lifetime management.)

Retry is implemented with an exponential backoff based on powers of 2
(2^0, 2^1, 2^2...) seconds. This is intended to help keep us from
exacerbating an issue if a remote resource really is down. By default,
we will only retry an operation three times before allowing an exception
to bubble up to the user. However, this value is defined in app
settings so that we can easily change it if needed.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/f52a498aab348cc86ba48b1aa3293aca1d493728