🔄 Wire up attribute-based interceptor for retry policy
Introduces an interceptor for classes which have the `RetryOnFault`
attribute, which will wrap their methods in a retry policy. Like our
transaction system, using interceptors for this saves us a lot of
boiler-plate code and helps ensure that retry policies are consistently
and correctly setup. However, it also introduce some challenges.

For example, we don't want to nest retry policies since that could
result in an exponentially-increasing number of retries based on the
level of nesting. A naive interceptor could make it really easy to
inadvertently nest retry polices and simultaneously make such nesting
difficult to discover. This could result in unexpected behavior when we
actually need to retry an operation. This change reduces that risk by
introducing a `RetryPolicyCoordinator` which keeps track of the number
of retry policies issued in the current lifetime (request) and ensures
that a real retry policy is only issued the first time one is requested;
all other retry policy requests receive a noop implementation.

Reasoning about how retry policy and transactions interact can be
difficult. Right now, the retry policy effectively lives "inside" the
transactions (based on the order the interceptors are wired up). This
means that transactions won't roll back due to an exception for which we
would want to retry, unless the maximum number of retries was exceeded.
This makes retries effectively invisible to other operations within the
same transaction (e.g. we don't have to rollback changes to neo4j
because of an intermittent network issue talking to SQL Server).

To accomplish the task, a couple of changes were made to the way
interceptors work generally. First, we were not correctly detecting
class attributes which were used to mark things we wanted to intercept.
This change corrects that, so that we can use these attributes on
classes or methods. Second, the `NonRacingAsyncInterceptor` base class
wasn't setup for us to easily get to the intercepted method; it only exposed
`beforeInvoke()` and `afterInvoke()` hooks. This didn't line up with the
way retry policy works which uses a single Execute method to wrap an
operation in the policy. This change updates that base class to allow
overriding of the `Invoke` method, which allows us to more easily wire
in the retry policy.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/dd77cf24d12ca699bb4c828e39060301adae03ec