🔄 Fix return value for retry policy interception around non-generic async methods
Previously, when intercepting non-generic async methods with the
RetryAsyncInterceptor, we were not setting the return value to the
resulting task from the intercepted method. This caused that task to
fall into a black hole, never to be seen again. This change sets the
return value on the invocation to the return value of the intercepted
method.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/6f77e769478c9e5679adb9e9437fac4ec0d1afe9