🔄 Fix ❌🤐🔙 Delete
❌🤐🔙 delete wasn't working because the interceptor module was detecting
that no `[FreezeTime]` attribute existed on either the method or the
class. It called '.SingleOrDefault()' on the method attributes, failed with
null, and then tried to call `.Single()` on the class attributes. This threw
an exception, as the class also had no `[FreezeTime]` attribute.

If we don't return a `TAttribute` out of this class (i.e. we return null),
we don't execute an interceptor. This is the desired behavior for this
case, as we don't care to `[FreezeTime]` on ❌🤐🔙 Delete.

Changing the `.Single()` call to `.SingleOrDefault()` will allow the
interceptor module to return null and not throw an exception.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/885e0fecc01a36c5ee679495597016962aa181d8