🔄 Fixes MVC Filter Redundant Execution
Fixes a problem wherein filter instances that are resolved more than once for a request (when they happen to implement more than one kind of filter interface) were getting executed redundantly. That is, they were getting executed at a multiplier equal to the number of filter interfaces they implemented, instead of only once per filter implementation.

This fixes the problem by wrapping the filters in simple facades that only expose the desired filter interface respective to the execution point in which they are resolved.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/dbab18cc77103ebe3722bc155930fb0a3fb3dac1