🔄 Ability to register the `SerilogLoggerFactory`
I needed to retrieve some configuration prior to registering the logger. `UseSerilog` is invoked before the `Startup` class. I sought to register the Factory manually but found out that is internal. I think that `SerilogLoggerFactory` is the main take away from this library and therefore exposing it would make it more flexible for those who don't want to use the extension methods.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/c2fee613a107a797cdfe326197e15ef9332f42d8