🔄 Converts Project Files and References to Target ASP.NET Core 2.2 on .NET Framework 4.7.2
Converts our projects to target either `netstandard2.0` or `net472`, as respectively required, in order for us to use `dotnet` tooling to build and deploy our websites targeting ASP.NET Core 2.2 instead of ASP.NET MVC 5. Specifically, we are reaching for ASP.NET Core 2.2 on .NET 4.7.2 because we don't want to replace our usage of Entity Framework 6 at this time, and EF6 won't run on the `netcoreapp` runtime.

In and of itself, this horribly breaks the build. We are attempting to skip over years of Microsoft refactoring the ASP.NET framework, and must contend with renamed libraries, changed namespaces, and in some cases even changed abstractions. This commit attempts to deal with project file and build configuration changes that are necessary for the upgrade, paving the way for future commits to deal with other classes of changes that need to be made in order for everything to work again.

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/fcb650e2565e13dcdff8dc9f83052bca624e05ab