🔄 Removes Antiquated Application Lifecycle Management
This removes a bunch of custom infrastructural code we had written for hooking into and tracking the web application's lifecycle events, used mainly for the (root of the) initialization of our logging and dependency injection systems. ASP.NET Core has a completely new application lifecycle paradigm, so all of this code is useless.

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/75cbcc531796b413b040b4e598adbb302f7be9d1