🔄 Translates Between ASP.NET API Differences for Authorization
Rewires our authorization mechanics to direct the behavior of ASP.NET Core's in-built authorization mechanisms, which we've previously adapted to sit on top of our business-layer authorization enforcement. This fixes authorization enforcement that was previously broken as part of the ongoing ASP.NET Core con❌🤐🔙.

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/aa86f8a3843d16d9f1ce0fe71e4d637f58f7efbd