🔄 Translates Between ASP.NET API Differences for Logging
This doubles-down on the idea that we don't need to be paranoid about whether injected enrichments are from singleton or request scope, which likely only mattered when we had scope-less logger creation in legacy ASP.NET (pre-Core).

This also carefully constrains our logging verbosity to make sure that future log-event-generators (such as the ASP.NET Core framework) do not log too noisily.

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/c5047cbb559372463d1c088745cfda5b861e802d