🔄 Adds Support for Singleton Injection of `IUrlHelper` on ASP.NET Core 2.2
For use in our AutoMapper mappings, introduces an implementation of `IUrlHelper` that is (nearly) safe to use from a singleton context after we update to ASP.NET Core.

NOTE: This adds new code that is not yet used or automatically tested. This new code will be used as of the coming ASP.NET Core upgrade, and has been manually tested to be working "well enough" for now.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/36efee4678c90921dcddae212c8db50966e864b8