🔄 Refactor custom interceptors to reduce thread contention
Previously, the `NonRacingAsyncInterceptor` had a mixture of sync and
async code which necessitated blocking `Wait` calls on threads (which
were taken from the same threadpool that web requests are spawned from),
which led to a lot of thread contention. This changes seeks to reduce
that thread contention by refactoring the interceptors to not use `Wait`
calls.

In the previous design, our custom interceptors would derive from
`NonRacingAsyncInterceptor` which would implement the `IInterceptor`
interface from Castle.DynamicProxy. Since Castle.DyamicProxy didn't
expose async methods, we didn't either, which led to the mixing of sync
and async code. In the new design, custom interceptors implement
`IPumpInterceptor` which exposes two methods: `Intercept` and
`InterceptAsync`. This allows custom interceptors to easily write async
code with proper `await` support. An adapter has been added that adapts
an instance of `IPumpInterceptor` to fit the `IInterceptor` interface.
The `AsyncAwareSwitchingInterceptor` has been removed in favor of
delegating to the appropriate method on `IPumpInterceptor` from the
above-mentioned `InterceptorAdapter`.

Secondary benefits of this change:
- the async implementations for the `ApplyTransActionInterceptor` and the
  `RetryInterceptor` have been greatly simplified since they can use `await`.
- the redundancy between the `NonRacingAsyncInterceptor` and the
  `RetryInterceptor` has been eliminated since interceptors can access
  the invocation.
- The need to have separate intercepors for sync vs async invocations
  has been removed.

Credit where credit is due: Bruce Markham drafted this new design and
Drew Tranquada did most of the work to implement it; I only carried the
baton for the last leg.

Co-authored-by: Drew Tranquada <dtranquada@foxguardsolutions.com>
Co-authored-by: Bruce Markham <bmarkham@foxguardsolutions.com>


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/b2e64f75cd14db672c91ba1c0889391a28379fd6