🔄 Adds Support for Idiomatic Configuration on ASP.NET Core 2.2 w/ AWS Elastic Beanstalk
This adds preliminary support for mitigating a [known deficiency](https://stackoverflow.com/questions/40127703/aws-elastic-beanstalk-environment-variables-in-asp-net-core-1-0) wherein AWS Elastic Beanstalk IIS Environment configuration doesn't actually get its configuration values into ASP.NET Core applications hosted on such. This mitigation is accomplished by implementing an (ASP.NET Core) idiomatic configuration provider that reads the AWS Elastic Beanstalk instance's well-known configuration file, and extracts the IIS information from it.

With this in place, it is possible for it to be consumed via an `AddAwsElasticBeanstalkIisEnvConfiguration` call while using an ASP.NET Core `IConfigurationBuilder` to build configuration, the same as with `AddJsonFile`, `AddIniFile`, etc.

NOTE: This adds new code that is not yet used or automatically tested. This new code will be used as of the coming ASP.NET Core upgrade, and has been manually tested to be working "well enough" for now.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/2df7324da3caf7a3cbeb87cd13f2d3201661c832