🔄 Translates Between ASP.NET API Differences for Test Helpers
This (starts to) fix tests that were broken as part of the con❌🤐🔙 to ASP.NET Core.

This change extracts Mvc-specific test helpers into dedicated libraries in order to help suppress future referential coupling problems as the Mvc-specific test helpers grow over time.

This adds helper to mock the current user, and uses it to fix a couple tests.

This adds a custom mock of `ITempDataDictionary` so that tests that depend on a pseudo-working implementation will be fixed.

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/394d1f3573bed69ce8a9170448defd627852d29e