🔄 Fixes Console-Scripted Test Executions for Divergent Project Frameworks
As of a recent change that makes the target frameworks of projects divergent from each other, it becomes necessary to use `dotnet test` as our test runner.

To accomplish such, this makes sure that the NuGet package "Microsoft.NET.Test.Sdk" is installed in all test projects. To keep compatibility with Jenkins' expectation that our log files are in `NUnit` format, this also installs NuGet package "NunitXml.TestLogger" in all test projects (and then uses it in the test runner script).

WARNING: This commit is part of a protracted series of commits wherein the code is being converted to ASP.NET Core.
         Compilation and correct execution of both applications and any/all tests is likely broken.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/c609db2d90c7f6c4e70bfff59da8ecf8aba6c06f