🔄 Updates All `TargetFramework`s (net471 -> net472)
This aligns us with the ❌🤐🔙 of the .NET Runtime that we intend to run the application on as of the upcoming upgrade to ASP.NET Core.

Also installs NuGet package `System.Net.Http`, since it no longer automatically resolves from the GAC after switching to `net472`.

This will allow us to test the runtime upgrade in isolation from the upcoming framework changes.

For more information, see: https://devblogs.microsoft.com/dotnet/announcing-the-net-framework-4-7-2/


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/19aaaaf9395b2ba2580e962707dd3e853cd032cd