🔄 Adds Support for Idiomatically Configuring Forwarded Headers on ASP.NET Core 2.2 w/ AWS ALBs
ASP.NET Core has support for using forwarded headers from reverse-proxies, however it can be tricky to configure correctly. This adds an adapter than configures the underlying support for forwarded headers, while expressing its API in terms of the configuration of an expected AWS Application Load Balancer as the reverse-proxy.

NOTE: This adds new code that is not yet used or automatically tested. This new code will be used as of the coming ASP.NET Core upgrade, and has been manually tested to be working "well enough" for now.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/ac02c18cddf40669b4ef88a8a519d6c7711a9601