🔄 Adds Support for Using Already-Registered Serilog Logger as ASP.NET Core's Internal Logger
This takes pre-written support for adapting Serilog to ASP.NET Core's logger abstraction, and adds an overload that makes it expect an already-registered instance of the Serilog logger. This will allow us to use both our legacy logging system (built directly on Serilog) alongside ASP.NET Core doing logging the way it wants to (which will be wired into Serilog).


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/a115fa4bcbbb9f75e576c7d7c9887e3288061b43