🔄 Implements Extensible Equivalent of `AutofacServiceProviderFactory`
When we upgrade to ASP.NET Core, we will need an easy way to idiomatically marry our custom paradigm for creating Autofac containers with the way ASP.NET Core expects DI containers to be created. This implements the `IServiceProviderFactory<…>` interface and provides a helper method to consume it, which is part of what ASP.NET Core will expect.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/31c049abc4707dde0f758daaed73b340eef35c68