🔄 Fixes Latent Static Code Analysis Violations
Historically our static code analysis configuration has not been consistent across projects. Before realigning the configurations we must correct the code to the configurations we wish to align to, so this does such.

Since the static analysis re-enablement (to follow this commit) is being done as part of the overall project con❌🤐🔙 effort that is happening as part of the ASP.NET Core con❌🤐🔙, rule violations that require undue refactoring right now have been suppressed in order to leave such refactorings as an exercise for future maintainers.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/23fe0197b4e8ef17675436978abf508edfabfa07