🔄 Re-enables & Re-configures Static Code Analysis
As part of the project con❌🤐🔙 as part of the upgrade to ASP.NET Core, it was necessary to temporarily disable static code analysis so as to minimize the compiler noise while chasing down API incompatibilities. Now that the codebase has been fitted into to the new API, this re-enables static code analysis and aligns the static analysis configuration across all projects in the solution.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/fd854d4032e7ed38698a9384d02f28c04366c943