🔄 Makes Entity Framework Health Checks Support Lazy Resolution
In some architectures (like PUMP), the act of attempting to resolve the Entity Framework data context can actually, itself, result in an exception.

This refactor allows resolution exceptions to occur as the failure of the specific health check rather than not getting caught, bubbling all the way out, and obscuring the other health checks.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/3c3d4e9b84dbe50a7786230aa65ba9da85aad737