🔄 Re-implements Server-side Validation for `RequiredIfTrue` Validation Attribute
This attribute performs server-side validation that a certain member is required, contingent on a second (boolean) member having a value of `true`.

This was originally provided to us via the "MVC Foolproof Validation" library, which we were unfortunately unable to keep when we transitioned to ASP.NET Core (because it doesn't support such.)

The implementation was written by taking `https://github.com/StefanKern/foolproof/blob/1996fa42dbbcad228c6bca7f790c905252bb4514/Foolproof/RequiredIf.cs` and unraveling the type hierarchy it is part of.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/5485ee414fc0e3fcf3fc10ca0fca45d44f6cece0