🔄 Re-implements Client-side Validation for `RequiredIfTrue` Validation Attribute
When we switched from ASP.NET Core, we had to remove the "MVC Foolproof Validation" because it did not support such. We recently re-implemented the server-side validation logic for `RequiredIfTrue`, but that wasn't enough to reinstate client-side validation.

Fortunately, we still have requisite scripts referenced by our client-side code for the respective validation, and it has merely been awaiting for us to annotate our HTML in the correct way to trigger it. This takes a Foolproof Validation's fork's ASP.NET Core compatible implementation of `IValidationAttributeAdapterProvider` and tools it to work with our new implementation of `RequiredIfTrue`. This means our emitted HTML is now properly annotated to trigger the client-side validation that, as noted above, lies in wait of such.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/de20ae8afd002a5740595d670956a139a32bd511