🔄 Setup Azure Pipelines job to publish NuGet packages for shared libraries
As PUMP has matured, we've developed several components that could be
generally useful in other similar projects. To facilitate code sharing,
we're going to create NuGet packages for those components. This will
allow us to easily import them into other projects and manage updates
the same way we do for 3rd party packages.

Azure DevOps gives us some artifact storage with our subscription, and
makes it trivial to setup a NuGet package feed. This integrates nicely
with Azure Pipelines, so this change adds an Azure Pipelines job
definition (instead of Jenkins like we would normally use) to create and
publish NuGet packages for components that are intended to be reusable.

We already have a robust build, test, and delivery pipeline for PUMP in
jenkins, so this job does not include any of the normal quality checks
(static analysis, tests, manual approvals, etc.). It runs on changes to
master that include files from the targeted components and only builds
and publishes the NuGet packages.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/9bd1e611faabbdfea90c3126b49888012ea8dee1