🔄 Add support for .NET Standard 2.1 to extensions projects
Configures extensions projects to target both .NET Standard 2.0 and .NET
Standard 2.1. This will start to pave the way for us to upgrade the web
projects to ASP.Net Core 3.0. This also makes it possible for us to use
the projects as NuGet packages for projects that are already on ASP.Net
Core 3.0 (see 9bd1e611).

Since Microsoft changed some APIs and reorganized namespaces in ASP.Net
Core 3.0, some of the ASP.Net Core specific projects had to be modified
to use conditional imports of the appropriate packages for 2.2 vs. 3.0.
Similarly, there were a couple of places where using statements needed
to be adjusted based on the target framework. However, they all were
simply nominal differences for our purposes, so we could just create an
alias that was switched based on the target framework.

Additionally, this change addresses a couple of warnings that popped up
as a result of expanding the framework support.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/b4f320f032f7860d1c3d1895e65f7dca9fe627a0