🔄 🏗🚛 Renames `FGS.Pump.Extensions.DI` -> `FGS.Autofac.Registration.Extensions`
This furthers an ongoing effort to break projects of helper functionality up into finely-grained units, in order to promote reusability along with transitive dependency stability. As part of such an effort, there is a renewed stringency on project naming necessary to convey a components purpose and pertinent primary dependency (in this case: `Autofac`).

This is a necessary prerequisite of making less-trivial PUMP functionality similarly reusable in the near future.


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/e621548c3e4987fa8a6e38d0869603de63e65fda