🔄 Fix `UseAutofacWithModulesProvider` for ASP.NET Core 3
ASP.NET Core changed the way hosting worked from 2.1. Our extension
methods for wiring in autofac had previously assumed that the change was
just a reorganization of the namespaces and that they could just
conditionally change some usings to be compatible with the new APIs.
However, the way that DI containers get wired in changed as well. This
meant that the process for setting up Autofac was different. This change
modifies the way that we wire in autofac when we're targeting .net
standard 2.1 or .net core 3.0 to correct this issue.

See https://autofaccn.readthedocs.io/en/latest/integration/aspnetcore.html#asp-net-core-3-0-and-generic-hosting


🔄 Original Commit: https://github.com/foxguardsolutions/Pump/commit/03c70f9a618ca3547ad90503290259a12ccd460a