👷✨ Adds Continuous Integration Builds
This makes sure that all code changes are buildable.

Note the inclusion of `dotnet pack`. Although it isn't the final `pack` to make packages for actual redistribution, it is instead included to make sure that _it will run successfully when the time comes_. (Our deployments are currently performed by the pipelines at https://dev.azure.com/foxguardsolutions/Foundation/.)

The other `dotnet` CLI subcommand executions are split out (despite being combinable to `dotnet test`) in order to give maximum visibility on workflow progress and results.

Windows has to be used for the build as long as we have `net472` as a `TargetFramework`.