👷‍♂️🐛 Corrects Iterative Insertion
Per [the documentation](https://docs.microsoft.com/en-us/azure/devops/pipelines/process/templates?view=azure-devops#iterative-insertion), a `sequence` is an acceptable type for the iterative insertion syntax, however a previous implementation has assumed that a `sequence`-typed child of `variables` would work as the input. Instead, the documentation demonstrates them as children of `parameters`, which may be an important detail. This switches the usage as our latest attempt to correct the error `A sequence was not expected`.