👷🐛 Makes Deployment Pipeline's `Deploy` Stage into Template
For a pipeline `stage` to be a template, (and therefore have `parameters` than can be used with iterative insertion), it must be in a file separate from the root of a "pipeline" and references as a template.

That is, our previous attempts to define `parameters` at the root of the pipeline YAML violates the schema because they are only valid for templates, and templates are only valid for `stage`s, `job`s, and `step`s.

For more information on the general pipeline schema, see: https://docs.microsoft.com/en-us/azure/devops/pipelines/yaml-schema?view=azure-devops&tabs=schema#pipeline
For more information on the template syntax, see: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/templates?view=azure-devops