👷🐛 Gives Valid Identifiers to `Deploy` Jobs
The names of Visual Studio projects, particularly when they are of "dotted identifier-dereferencing" form, are not compatible with the naming constraints of jobs in Azure Pipelines. This abandons the idea of using the literal project name as part of the job name, and instead uses an ordinated numeral.

While it is expected that this will be fragile to manually maintain, we favor this more than the only other obvious & available approach at the moment (which is to push the expansion of the list down into a variable number of steps for a single job, which would then be forced to run in sequence and making it incredibly fragile to transient build problems.)

For more information about functions available to expressions in Azure Pipelines, see: https://docs.microsoft.com/en-us/azure/devops/pipelines/process/expressions?view=azure-devops#functions