♻🐛 Fixes Broken Package Separation for `FGS.*.Logging.Serilog`
A previous refactoring attempted to split this library in two by varying the assembly name based on the target framework. Unfortunately, this did not have the intended effect of varying the output package name (which is documented as defaulting to the assembly name.) Additionally, this also had the misfortune of making the NuGet Package Manager tool in Visual Studio become unstable and crash when attempting to manage dependencies elsewhere in the parent SLN.

This fixes the attempted split by duplicating the project and then specializing each copy for the target frameworks it is intended for.

This also deletes the associated test project because it contained nothing substantive and thus was not worth duplicating at this time.