✨ Adds Query Projection Extension Method
This adds `Map` extension methods to `IQueryable`-based queries.

`Map` is a white-labelling facade on top of `AutoMapper`'s `ProjectTo` functionality, but has the added benefit of being stable (in name and signature) across AutoMapper versions. As such, the functionality's implementation is separately distributed between AutoMapper 7 and newer versions of AutoMapper.