➕ Adds Project for Documentation Builds
This adds a project that serves as a surrogate host for the `DocFx` documentation builder. The project is purposefully not added to the `sln`, as it is not a requirement of `DocFx` to do so and adding it would slow development builds.

For more information, see: https://dotnet.github.io/docfx/tutorial/docfx_getting_started.html#3-use-docfx-integrated-with-visual-studio